Reorganização das pastas de crypto
