+++ src/models/crypto/enumerator.rs
@@ -1,19 +1,19 @@
 use serde::{Deserialize, Serialize};
-
-use super::{salting::SaltingStrategy, util::EnumStringParse};
-
-#[derive(Debug, Clone, Deserialize, Serialize, Queryable, Default)]
-pub struct PlainTextAlgorithmn;
+use crate::models::crypto::{ApplyCripto, plain::*, sha1::*};
+use crate::models::salting::SaltingStrategy;
+use crate::models::util::EnumStringParse;
 
 #[derive(Debug, Clone, Deserialize, Serialize)]
 pub enum CryptoAlgorithm {
     Plain(PlainTextAlgorithmn),
+    Sha1(Sha1Algorithm)
 }
 
 impl CryptoAlgorithm {
     pub fn apply(&self, key: &str, salt: &Vec<&str>, salter: &SaltingStrategy) -> String {
         match self {
             CryptoAlgorithm::Plain(c) => c.apply(key, salt, &salter),
+            CryptoAlgorithm::Sha1(c) => c.apply(key, salt, &salter)
         }
     }
 }
@@ -28,17 +28,8 @@ impl EnumStringParse for CryptoAlgorithm {
     fn str_to_enum(key: &str) -> Option<Self> {
         match key {
             "plain" => Some(CryptoAlgorithm::Plain(PlainTextAlgorithmn::default())),
+            "sha1" => Some(CryptoAlgorithm::Sha1(Sha1Algorithm::default())),
             _ => Option::None,
         }
-    }
-}
-
-trait ApplyCripto {
-    fn apply(&self, key: &str, salt: &Vec<&str>, salter: &SaltingStrategy) -> String;
-}
-
-impl ApplyCripto for PlainTextAlgorithmn {
-    fn apply(&self, key: &str, salt: &Vec<&str>, salter: &SaltingStrategy) -> String {
-        salter.apply(key, &salt)
     }
 }
